backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
httpx>=0.26.0
celery[redis]>=5.3.0
redis>=5.0.0
email-validator>=2.0.0
aiosmtplib>=3.0.0
cryptography>=42.0.0
